refactor(secrets): Move `getEnumOrDefault()` to `config-utils`
Add documentation along the way as the function is now public.
